feat(codec): added in a column for video codec
Added in the video codec because that can also be key to figuring out if
something is encoded at too high of a bitrate. A video can appear
encoded at too high of a bitrate if it is hevc but not if it is h264,
because the latter is a less-efficient codec.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>